+++ src/main/java/handler/auth/LoginHandler.java
@@ -1,4 +1,4 @@
-package handler;
+package handler.auth;
 
 import static webserver.common.Constants.BLANK;
 import static webserver.common.Constants.EMPTY;
@@ -10,6 +10,7 @@ import static webserver.common.Constants.SLASH;
 import static webserver.common.Constants.USER_ID;
 
 import db.Database;
+import handler.Handler;
 import model.User;
 import webserver.common.HttpStatus;
 import webserver.exception.NotRegisteredUserException;
@@ -17,7 +18,7 @@ import webserver.request.Request;
 import webserver.response.Response;
 import webserver.session.Session;
 
-public class LoginHandler implements Handler{
+public class LoginHandler implements Handler {
     private static final String NOT_REGISTERED_USER = "아이디 또는 비밀번호가 올바르지 않습니다.";
     private static final String ROOT_PATH = "Path=/";
     @Override
