+++ src/main/java/handler/util/PathResolvedHandler.java
@@ -1,5 +1,6 @@
-package handler;
+package handler.util;
 
+import handler.Handler;
 import java.util.Map;
 import webserver.request.Request;
 import webserver.response.Response;
