+++ src/main/java/handler/view/MainPageHandler.java
@@ -1,9 +1,10 @@
-package handler;
+package handler.view;
 
 import static webserver.common.Constants.EMPTY;
 import static webserver.common.Constants.SLASH;
 
 import db.Database;
+import handler.Handler;
 import java.util.List;
 import java.util.Optional;
 import model.Article;
@@ -15,30 +16,29 @@ import webserver.loader.ResourceLoader;
 import webserver.request.Request;
 import webserver.response.Response;
 
-public class ArticleDetailHandler implements Handler{
+public class MainPageHandler implements Handler {
     private final ResourceLoader resourceLoader;
 
-    public ArticleDetailHandler(ResourceLoader resourceLoader) {
+    public MainPageHandler(ResourceLoader resourceLoader) {
         this.resourceLoader = resourceLoader;
     }
 
     @Override
     public Response handle(Request request) {
-        int id = Integer.parseInt(request.getPathVariables().get("id"));
-        Article article = Database.findArticleById(id);
-        Optional<Article> prev = Database.findPreviousArticle(id);
-        Optional<Article> next = Database.findNextArticle(id);
+        Article article = Database.findAllArticles().reversed().getFirst();
+        Optional<Article> prev = Database.findPreviousArticle(article.getId());
+        Optional<Article> next = Database.findNextArticle(article.getId());
 
-        byte[] template = resourceLoader.fileToBytes(SLASH, true);
+        byte[] html = resourceLoader.fileToBytes(SLASH, true);
         List<TemplateRenderer> renderers = List.of(
                 new HeaderRenderer(),
                 new ArticleContentRenderer(article, prev, next)
         );
 
         for (TemplateRenderer renderer : renderers) {
-            template = renderer.render(null, template); // 유저는 필요하면 세션에서
+            html = renderer.render(null, html);
         }
 
-        return new Response(HttpStatus.OK, template, EMPTY);
+        return new Response(HttpStatus.OK, html, EMPTY);
     }
 }
