+++ src/main/java/webserver/resolver/MethodResolver.java
@@ -1,15 +1,15 @@
 package webserver.resolver;
 
-import handler.ArticleDetailHandler;
-import handler.CreateUserHandler;
+import handler.article.ArticleDetailHandler;
+import handler.auth.CreateUserHandler;
 import handler.Handler;
-import handler.LoginHandler;
-import handler.LogoutHandler;
-import handler.MainPageHandler;
-import handler.PathResolvedHandler;
+import handler.auth.LoginHandler;
+import handler.auth.LogoutHandler;
+import handler.view.MainPageHandler;
+import handler.util.PathResolvedHandler;
 import handler.StaticFileHandler;
-import handler.TemplateHandler;
-import handler.WriteArticleHandler;
+import handler.view.TemplateHandler;
+import handler.article.WriteArticleHandler;
 import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.HashMap;
@@ -19,10 +19,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import template.ArticleContentRenderer;
 import template.HeaderRenderer;
 import template.UserListRenderer;
-import webserver.WebServer;
 import webserver.common.HttpMethod;
 import webserver.exception.MethodNotAllowedException;
 import webserver.loader.FileResourceLoader;
