+++ src/test/java/handler/LoginHandlerTest.java
@@ -4,8 +4,8 @@ package handler;
 import static org.assertj.core.api.Assertions.*;
 
 import db.Database;
+import handler.auth.LoginHandler;
 import java.lang.reflect.Field;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
