+++ src/test/java/handler/LogoutHandlerTest.java
@@ -2,6 +2,7 @@ package handler;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import handler.auth.LogoutHandler;
 import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.List;
